Do not update_option for DB version on every request
The install_db function is hooked into the plugins loaded action, but includes an `update_option` call which is fired even when the stored option DOES match.

This is inefficient, and can cause unnecessary DB write load. It is especially problematic on some enterprise hosting where an UPDATE call causes all subsequent reads in a request to go through a Writer database instead of a read replica, overwhelming production DB servers.